requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Django==5.2.3
django-ninja==1.4.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
lxml==5.4.0
packaging==25.0
platformdirs==4.3.8
pydantic==2.11.5
pydantic_core==2.33.2
pytz==2025.2
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
zeep==4.3.1
